+++ src/either.ts
@@ -1,10 +1,12 @@
 import type { Monad, Comonad } from './types';
 
 class EitherConstructor<L extends Error, R> implements Comonad, Monad {
-  readonly #value: L | R;
+  // TODO: review this when ECMAScript's private class fields will be
+  // widely spread in browsers.
+  private readonly _value: L | R;
 
   private constructor(value: L | R) {
-    this.#value = value;
+    this._value = value;
   }
 
   /** Creates `Either` monad instance with **Left** state. */
@@ -22,9 +24,7 @@ class EitherConstructor<L extends Error, R> implements Comonad, Monad {
    * (**Right** or **Left**) of `Either` by yourself.
    * If value is `Either`, then its copy will be returned.
    */
-  static either<A extends Error, B>(
-    value: A | B | Either<A, B>
-  ): Either<A, B> {
+  static either<A extends Error, B>(value: A | B | Either<A, B>): Either<A, B> {
     const exposedValue = isEither<A, B>(value) ? value.extract() : value;
     return exposedValue instanceof Error
       ? EitherConstructor.left<A, B>(exposedValue)
@@ -36,7 +36,7 @@ class EitherConstructor<L extends Error, R> implements Comonad, Monad {
   }
 
   isLeft(): this is Either<L, never> {
-    return this.#value instanceof Error;
+    return this._value instanceof Error;
   }
 
   map<A>(fn: (value: R) => L | A): Either<L, A> {
@@ -46,15 +46,15 @@ class EitherConstructor<L extends Error, R> implements Comonad, Monad {
   /** Maps inner value if it is not an `Error` instance. Same as `Either.map`. */
   mapRight<A>(fn: (value: R) => L | A): Either<L, A> {
     return this.isRight()
-      ? EitherConstructor.either<L, A>(fn(this.#value))
-      : EitherConstructor.left<L, A>(this.#value as L);
+      ? EitherConstructor.either<L, A>(fn(this._value))
+      : EitherConstructor.left<L, A>(this._value as L);
   }
 
   /** Maps inner value if it is an `Error` instance */
   mapLeft<E extends Error>(fn: (value: L) => E | R): Either<E, R> {
     return this.isRight()
-      ? EitherConstructor.right<E, R>(this.#value)
-      : EitherConstructor.either<E, R>(fn(this.#value as L));
+      ? EitherConstructor.right<E, R>(this._value)
+      : EitherConstructor.either<E, R>(fn(this._value as L));
   }
 
   apply<U>(other: Either<L, (value: R) => L | U>): Either<L, U> {
@@ -65,12 +65,12 @@ class EitherConstructor<L extends Error, R> implements Comonad, Monad {
 
   chain<U>(fn: (value: R) => Either<L, U>): Either<L, U> {
     return this.isRight()
-      ? fn(this.#value)
-      : EitherConstructor.left<L, U>(this.#value as L);
+      ? fn(this._value)
+      : EitherConstructor.left<L, U>(this._value as L);
   }
 
   extract(): L | R {
-    return this.#value;
+    return this._value;
   }
 }